pyproject.toml
[project]
name = "new-me"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "greenlet>=3.0.0",
    "apscheduler>=3.10.4",
    "aiohttp>=3.11.11",
    "gTTS>=2.5.0",
]

[tool.black]
line-length = 110
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 110
src_paths = ["src"]

[tool.ruff]
line-length = 110
target-version = "py313"
select = ["E","F","I","B","UP"]
ignore = []
fix = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
